scripts/image_collector/requirements.txt
# pip install -r requirements.txt

httpx>=0.27.0
rembg>=2.0.50
onnxruntime>=1.16.0
Pillow>=10.0.0
